Add item scripts, item types and on destroy particle
- Can choose type for items: restore health for small amount or full, and increase stat move speed, max health, damage
- Add particle prefab when pick up items

+++ Unity Project/Math Conquest 3D (Core)/Assets/Script/mode exploration script/EM player system/Player Script/ExplorationModePlayerHealth.cs
@@ -71,7 +71,7 @@ public class ExplorationModePlayerHealth : MonoBehaviour
         playerHealthCurrent = 0;
     }
 
-    public void PlayerStoreHealth(int healthRestoreValue)
+    public void PlayerStoreHealth(float healthRestoreValue)
     {
         PlayerHealthBarControl();
     }
